长按手势AllowableMovement类型修改为double

+++ frameworks/bridge/declarative_frontend/engine/jsi/nativeModule/arkts_native_common_bridge.h
@@ -325,7 +325,7 @@ public:
     static void GetTapGestureValue(ArkUIRuntimeCallInfo* runtimeCallInfo, int32_t& fingers,
         int32_t& count, double& distanceThreshold, bool& limitFingerCount, uint32_t argNumber);
     static void GetLongPressGestureValue(ArkUIRuntimeCallInfo* runtimeCallInfo, int32_t& fingers, bool& repeat,
-        int32_t& duration, bool& limitFingerCount, int32_t& allowableMovement, uint32_t argNumber);
+        int32_t& duration, bool& limitFingerCount, double& allowableMovement, uint32_t argNumber);
     static void GetPanGestureValue(ArkUIRuntimeCallInfo* runtimeCallInfo, int32_t& fingers, int32_t& direction,
         PanDistanceMapDimension& distanceMap, bool& limitFingerCount, uint32_t argNumber);
     static Local<panda::ObjectRef> CreateTapGestureLocationInfo(
